requirements.txt
pandas==0.25.3
scikit-learn==0.22
feather-format
